.claude/skills/video-editor/requirements.txt
opentimelineio>=0.15.0
pydantic>=2.0.0
requests>=2.31.0
pyyaml>=6.0.0
